requirements.txt
discord.py>=2.4.0
google-genai>=0.5.0
pydub>=0.25.1
requests>=2.31.0
Telethon>=1.36.0
yt-dlp>=2024.5.27
uvloop>=0.19.0
pytest
flake8
black
isort
